Move utils/request into its own module
