+++ packages/client-library-otel/src/utils/request/index.ts
@@ -1,0 +1,6 @@
+export {
+  getRootRequestAttributes,
+  getRequestAttributes,
+  getResponseAttributes,
+  cloneRequestForAttributes,
+} from "./request";
